Allow adding fact without semicolon if body only contains attachments
Current behavior is fact_add() will pitch a fit if a user tries to add a fact that is only an attachment without including the semicolon. It would make sense to make an exception in this case and let the command go through anyway, because the message contains a valid fact without text in the fact body.